Create README.md
The Campus Requirement API Documentation is a Spring Boot REST API project designed to facilitate the management of campus job posting and  student applications. It allows institutions to post job opportunities and enables students to apply for these jobs. The primary entities in the system are Students, users, job Postings, Job Application and Interviews, The application aims to streamline the hiring process, improve communication between students and employers, and maintain an organized database opportunities.